Check if font exists before installing.

+++ install_scripts/install_dotfiles/fonts.sh
@@ -6,20 +6,17 @@ sudo apt install \
   fonts-font-awesome \
   dtrx
 
-# fonts_dir="$HOME/.fonts"
-# if [ ! -d "$fonts_dir" ]; then
-  # mkdir "$fonts_dir"
-  # chown -R $USER:$(id -g -n $USER) "fonts_dir"
-  # echo "~/.fonts dir created."
-  # else
-    # echo "Skip making ~/.fonts dir, it already exists"
-# fi
+spacemono_dir="$HOME/.local/share/fonts/SpaceMono"
+if [ ! -d "$spacemono_dir" ]; then
+    wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.0.0/SpaceMono.zip -P "$spacemono_dir"
+    cd "$spacemono_dir"
+    dtrx SpaceMono.zip
+    fc-cache -fv
+    rm SpaceMono.zip
+    cd
+    echo "SpaceMono font installed to $spacemono_dir"
 
-fonts_dir="$HOME/.local/share/fonts"
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.0.0/SpaceMono.zip -P "$fonts_dir"
-cd "$fonts_dir"
-dtrx SpaceMono.zip
-fc-cache -fv
-rm SpaceMono.zip
-cd
-echo "SpaceMono font installed to $fonts_dir"
+    else
+        echo "Skipping - SpaceMono font already installed"
+fi
+
